last commit

+++ AccountService/AccountService/Dtos/Response/UpdateBalanceResponse.cs
@@ -4,5 +4,5 @@ public class UpdateBalanceResponse
 {
     public bool IsSuccess { get; set; }
     public string Message { get; set; }
-    public double NewBalance { get; set; }
+    public decimal NewBalance { get; set; }
 }
